requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.2
Django==5.2
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
Markdown==3.7
pillow==11.2.1
pycparser==2.22
sqlparse==0.5.3
tzdata==2025.2
